feat: added ice breaker component
