+++ src/lang/en.json
@@ -7,7 +7,7 @@
   "splash.board_name": "Board Name",
   "splash.board_name_example": "Sprint 21 Retrospective",
   "splash.template": "Template",
-  "splash.icebreaking": "Ice Breaking",
+  "splash.icebreaking": "Ice Breaker",
   "splash.icebreaking_example": "If you could have a superpower overnight. What it would be?",
   "splash.create": "Start A Retro!",
   "splash.your_boards": "past boards",
